refactor(schemas): move remaining schemas into domain folders
- ScriptLanguageSchema → common/
- LogLevelSchema → log/
- RecipeSchema → recipe/
- SettingsSchema → settings/
- SkillSchema → skill/
- Update cross-folder import paths
