+++ packages/schemas/src/index.ts
@@ -4,12 +4,12 @@ export * from "./agent-trace";
 export * from "./distillation";
 export * from "./github-project";
 export * from "./job";
-export * from "./LogLevelSchema";
+export * from "./common";
+export * from "./log";
 export * from "./meta";
 export * from "./operation";
-export * from "./RecipeSchema";
+export * from "./recipe";
 export * from "./refinement";
-export * from "./ScriptLanguageSchema";
-export * from "./SettingsSchema";
-export * from "./SkillSchema";
+export * from "./settings";
+export * from "./skill";
 export * from "./pipeline";
