+++ packages/schemas/src/settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./SettingsSchema";
